move mysql queries to file sql.py in a queries folder
